Add classes to support an OO refactor
The code that existed was a mess, with lots of functions that use global
variables and functions contained in files with other unrelated code.

To tidy this up and make the code more maintainable and readable, an
object oriented approach is being adopted that will require a
significant refactor of the code base. A number of classes have been
created to support this, they are currently contained in the
"structure.js" file, however, this file is a temporary home and they
will be moved elsewhere later on.