Add a range for classes for OO
An AnalyticSolution class has been created. From this class many others
are derived, including RepeatedRootSolution (for solutions with repeated
roots) and ComplexRootSolutions (for solutions with complex roots).
These classes handle the formatting of LaTeX for printing the equation of
the solution to the screen, and time series generation for generating
data to plot.

A DifferentialEquation class has been created. Derived classes include
SecondOrderDifferentialEquation and FirstOrderDifferentialEquation.
These control finding and storing the roots of the equation and
formatting the LaTeX for printing the equation to the screen.

Two factory classes have been added for creating DifferentialEquation
and AnalyticSolution derived objects, with the exact concrete object
created depending on the settings provided.

Tests have been created in the test directory for these classes.